DoodleViewController 에서 DataSource 부분 분리
DataSource 생성시점에서 JSON 디코드가 완료되면 completion 을 통해 Notification 을 보내도록 구현
DoodleViewController 에서 Notification 을 받아 콜렉션 뷰를 리로드해줌
